increase max number of http request and response headers
The current limit of 32 https headers can be exceeded by some clients.
Between built-in headers from browsers and custom headers that may be
added in different places like CloudFlare or Nginx, this limit can be
hit by many clients. This commit increases the limits to 64, which is
much less likely to be reached.